Reorganize server.js, implement bare minimum for Apollo Server.
