+++ server/schemas/resolvers.js
@@ -1,0 +1,11 @@
+const resolvers = {
+  Query: {
+    placeholder: async () => {
+      return {
+        _id: "placeholder ID"
+      };
+    }
+  }
+}
+
+module.exports = resolvers;
