+++ server/server.js
@@ -1,11 +1,25 @@
 const path = require('path');
+const http = require('http');
 const express = require('express');
-const { Server } = require('socket.io');
+const { ApolloServer } = require('apollo-server-express');
+const { typeDefs, resolvers } = require('./schemas');
+const { Server: SocketIOServer } = require('socket.io');
+const registerGameHandlers = require('./utils/gameHandlers');
 const cors = require('cors');
 const db = require('./config/connection');
 
+// -------------------------------
+// Basic Express Setup
+// -------------------------------
+
 const PORT = process.env.PORT || 3001;
 const app = express();
+app.use(express.urlencoded({ extended: false }));
+app.use(express.json());
+// Rather than using "app.listen" (which returns an HttpServer object),
+// we create one explicitly here, and can use "server.listen" later
+// to start the server.
+const server = http.createServer(app);
 
 // Need to use CORS middleware during development
 // so that our socket server (on port 3001) can communicate
@@ -21,86 +35,55 @@ if (process.env.NODE_ENV === 'production') {
   app.use(cors());
 }
 
+// -------------------------------
+// Socket.io Setup
+// -------------------------------
 
-db.once('open', () => {
-  // This time, we actually need to use the HTTP server instance that is
-  // returned by 'app.listen', so we capture it in a 'server' variable.
-  const server = app.listen(PORT, () => {
-    console.log(`Server listening on port ${PORT}! 🧩`)
-  });
-
-  // Here we attach the socket.io server object (which will listen for socket events)
-  // to the HTTP server created by Express in the previous block.
-  // Again, need to set up CORS for a development environment.
-  const io = (process.env.NODE_ENV === 'production') 
-  ? 
-  new Server(server)
-  :
-  new Server(server, {
-    cors: {
-      origin: "http://localhost:3000",
-      methods: ["GET", "POST"]
-    }
-  });
-
-  // Listen for socket events.
-  io.on('connection', (socket) => {
-    console.log(`User Connected: ${socket.id}`);
-
-    socket.on("join_room", async (data) => {
-      await socket.join(data.roomId);
-      socket.data.playerName = data.playerName;
-      socket.data.isHost = data.isHost;
-      const roommateSockets = await io.in(data.roomId).fetchSockets();
-      const players = [];
-      for (roommate of roommateSockets) {
-        players.push({
-          id: roommate.id,
-          name: roommate.data.playerName,
-          isHost: roommate.data.isHost,
-          boardPosition: 0
-        });
-      }
-      console.log(`User ${socket.data.playerName} (${socket.id}) joined room ${data.roomId}.`);
-      console.log(players.length);
-      io.sockets.in(data.roomId).emit('receive_current_players',players);
-    });
-
-    socket.on("leave_room", async (data) => {
-      await socket.leave(data);
-      const roommateSockets = await io.in(data.roomId).fetchSockets();
-      const players = [];
-      for (roommate of roommateSockets) {
-        players.push({
-          id: roommate.id,
-          name: roommate.data.playerName,
-          isHost: roommate.data.isHost,
-          boardPosition: 0
-        });
-      }
-      console.log(`User ${socket.data.playerName} (${socket.id}) left room ${data}.`);
-      console.log(players.length);
-      io.sockets.in(data).emit('receive_current_players',players);
-    });
-
-    socket.on("host_left", (room) => {
-      socket.to(room).emit('receive_host_left');
-    });
-
-    socket.on("start_game", (room) => {
-      io.sockets.in(room).emit('receive_start_game');
-    });
-
-    socket.on("advance_turn", (room) => {
-      io.sockets.in(room).emit('receive_advance_turn');
-    });
-    
-    socket.on("send_message", (data) => {
-      socket.broadcast.emit("receive_message", data);
-    })
-
-    socket.on("disconnect", (reason) => {
-      console.log(`User ${socket.id} disconnected - ${reason}`);
-    });
-  });
+// Create a new socket.io server instance (which will listen for socket events).
+// Again, need to set up CORS for a development environment.
+const io = (process.env.NODE_ENV === 'production') 
+? 
+new SocketIOServer(server)
+:
+new SocketIOServer(server, {
+  cors: {
+    origin: "http://localhost:3000",
+    methods: ["GET", "POST"]
+  }
 });
+
+// Register socket events to listen for.
+io.on('connection', (socket) => {
+  console.log(`User Connected: ${socket.id}`);
+  registerGameHandlers(io, socket);
+});
+
+// -------------------------------
+// Apollo Server Setup
+// -------------------------------
+
+// Create an Apollo server instance.
+const apollo = new ApolloServer({
+  typeDefs,
+  resolvers,
+  // context: authMiddleware
+});
+
+// -------------------------------
+// Starting the server
+// -------------------------------
+
+const startServer = async () => {
+  await apollo.start();
+  apollo.applyMiddleware({ app });
+
+  db.once('open', () => {
+    server.listen(PORT, () => {
+      console.log(`Server listening on port ${PORT}! 🧩`);
+      console.log(`Use GraphQL at http://localhost:${PORT}${apollo.graphqlPath}`);
+    })
+  })
+};
+
+// Start the server.
+startServer();
